+++ R/SAGE-conditional.R
@@ -45,7 +45,7 @@ ConditionalSAGE = R6Class(
 
     #' @description
     #' Creates a new instance of the ConditionalSAGE class.
-    #' @param task,learner,measure,resampling,features,estimator,n_permutations,n_coalitions,batch_size,n_samples,early_stopping,se_threshold,min_permutations,check_interval Passed to [SAGE].
+    #' @param task,learner,measure,resampling,features,estimator,n_permutations,n_coalitions,max_features,batch_size,n_samples,early_stopping,se_threshold,min_permutations,check_interval Passed to [SAGE].
     #' @param sampler ([ConditionalSampler]) Optional custom sampler. Defaults to [ConditionalARFSampler].
     initialize = function(
       task,
@@ -53,9 +53,10 @@ ConditionalSAGE = R6Class(
       measure = NULL,
       resampling = NULL,
       features = NULL,
-      estimator = c("permutation", "kernel"),
+      estimator = c("permutation", "kernel", "exact"),
       n_permutations = NULL,
       n_coalitions = NULL,
+      max_features = 12L,
       sampler = NULL,
       batch_size = 5000L,
       n_samples = 100L,
@@ -85,6 +86,7 @@ ConditionalSAGE = R6Class(
         estimator = estimator,
         n_permutations = n_permutations,
         n_coalitions = n_coalitions,
+        max_features = max_features,
         batch_size = batch_size,
         n_samples = n_samples,
         early_stopping = early_stopping,