+++ R/SAGE-marginal.R
@@ -23,16 +23,17 @@ MarginalSAGE = R6Class(
   public = list(
     #' @description
     #' Creates a new instance of the MarginalSAGE class.
-    #' @param task,learner,measure,resampling,features,estimator,n_permutations,n_coalitions,batch_size,n_samples,early_stopping,se_threshold,min_permutations,check_interval Passed to [SAGE].
+    #' @param task,learner,measure,resampling,features,estimator,n_permutations,n_coalitions,max_features,batch_size,n_samples,early_stopping,se_threshold,min_permutations,check_interval Passed to [SAGE].
     initialize = function(
       task,
       learner,
       measure = NULL,
       resampling = NULL,
       features = NULL,
-      estimator = c("permutation", "kernel"),
+      estimator = c("permutation", "kernel", "exact"),
       n_permutations = NULL,
       n_coalitions = NULL,
+      max_features = 12L,
       batch_size = 5000L,
       n_samples = 100L,
       early_stopping = FALSE,
@@ -50,6 +51,7 @@ MarginalSAGE = R6Class(
         estimator = estimator,
         n_permutations = n_permutations,
         n_coalitions = n_coalitions,
+        max_features = max_features,
         batch_size = batch_size,
         n_samples = n_samples,
         early_stopping = early_stopping,